requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.1
catboost==0.24.4
cycler==0.10.0
dj-database-url==0.5.0
django-environ==0.4.5
django==3.1.5
graphviz==0.16
gunicorn==20.0.4
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.5
pandas==1.2.0
pillow==8.1.0; python_version >= '3.6'
plotly==4.14.3
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.5
retrying==1.3.3
scipy==1.6.0
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
